extract optional_time_delta for sharing
